+++ R/XenaQuery.R
@@ -14,6 +14,20 @@ XenaQuery <- function(x) {
   use_hiplot <- getOption("use_hiplot", default = FALSE)
 
   data_list <- UCSCXenaTools::XenaData
+
+  if (use_hiplot) {
+      # Check website status
+      use_hiplot <- tryCatch(
+          {
+              httr::http_error("https://xena-ucscpublic.hiplot.com.cn")
+              TRUE
+          },
+          error = function(e) {
+              message("The hiplot server may down, we will not use it for now.")
+              FALSE
+          }
+      )
+  }
   if (use_hiplot) {
     message("Use hiplot server (China) for mirrored data hubs (set 'options(use_hiplot = FALSE)' to disable it)")
     data_list$XenaHosts <- .xena_mirror_map_rv[data_list$XenaHosts]